chore: relocate the pack-4 seam tests to tests/ and speak the merged receipt
The rebase stranded automation-create.test.ts and make-tool.test.ts at the
old src/ paths with dead imports; they move to the per-package tests/
directory (#1081 layout) with imports mapped to the server/ structure
(engineOverAdapter seeding, contract-local ScreenAssembler). Assertions
follow the merged receipt semantics: partial status + the didn't-get-built
sentence, and the CreateServerWork envelope shape end to end.
